Move `wait_for_invocation_and_jobs()` method to `BaseWorkflowPopulator`
Remove redundant `wait_for_invocation()` call inside it.
